requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
frozenlist==1.3.3
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.23.3
ipython==8.12.2
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.4
openpyxl==3.1.2
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.8.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
pygame==2.5.0
Pygments==2.15.1
python-dateutil==2.8.2
python-weather==1.0.3
pywin32==306
pyzmq==25.1.0
requests==2.31.0
rsa==4.9
six==1.16.0
soundfile==0.12.1
srt==3.5.3
stack-data==0.6.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
urllib3==1.26.16
vosk==0.3.45
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
google-cloud-texttospeech